+++ share/pvss/pvss_test.go
@@ -12,15 +12,15 @@ import (
 
 func TestComputePolyCommitments(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
-	n := 20
-	t := 15
+	n := uint32(20)
+	t := uint32(15)
 	H := suite.Point().Pick(suite.XOF([]byte("H")))
 	secret := suite.Scalar().Pick(suite.RandomStream())
 	priPoly := share.NewPriPoly(suite, t, secret, suite.RandomStream())
 
 	x := make([]kyber.Scalar, n) // trustee private keys
 	X := make([]kyber.Point, n)  // trustee public keys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		x[i] = suite.Scalar().Pick(suite.RandomStream())
 		X[i] = suite.Point().Mul(x[i], nil)
 	}
@@ -33,7 +33,7 @@ func TestComputePolyCommitments(test *testing.T) {
 	indices := make([]uint32, n)
 	values := make([]kyber.Scalar, n)
 	HS := make([]kyber.Point, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		indices[i] = priShares[i].I
 		values[i] = priShares[i].V
 		HS[i] = H
@@ -43,12 +43,12 @@ func TestComputePolyCommitments(test *testing.T) {
 	require.NoError(test, err)
 
 	_, com := pubPoly.Info()
-	actualComm := computeCommitments(suite, n, com)
+	actualComm := computeCommitments(suite, int(n), com)
 
-	require.Equal(test, n, len(expectedComm))
+	require.Equal(test, n, uint32(len(expectedComm)))
 	require.Equal(test, len(expectedComm), len(actualComm))
 
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		require.Equal(test, expectedComm[i].String(), actualComm[i].String())
 	}
 }
@@ -57,11 +57,11 @@ func TestPVSS(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
 	G := suite.Point().Base()
 	H := suite.Point().Pick(suite.XOF([]byte("H")))
-	n := 10
+	n := uint32(10)
 	t := 2*n/3 + 1
 	x := make([]kyber.Scalar, n) // trustee private keys
 	X := make([]kyber.Point, n)  // trustee public keys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		x[i] = suite.Scalar().Pick(suite.RandomStream())
 		X[i] = suite.Point().Mul(x[i], nil)
 	}
@@ -75,7 +75,7 @@ func TestPVSS(test *testing.T) {
 
 	// (2) Share decryption (trustees)
 	sH := make([]kyber.Point, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		sH[i] = pubPoly.Eval(encShares[i].S.I).V
 	}
 
@@ -83,10 +83,10 @@ func TestPVSS(test *testing.T) {
 	var E []*PubVerShare // good encrypted shares
 	var D []*PubVerShare // good decrypted shares
 
-	globalChallenge, err := computeGlobalChallenge(suite, n, pubPoly, encShares)
+	globalChallenge, err := computeGlobalChallenge(suite, int(n), pubPoly, encShares)
 	require.NoError(test, err)
 
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		if ds, err := DecShare(suite, H, X[i], sH[i], x[i], globalChallenge, encShares[i]); err == nil {
 			K = append(K, X[i])
 			E = append(E, encShares[i])
@@ -104,11 +104,11 @@ func TestPVSSDelete(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
 	G := suite.Point().Base()
 	H := suite.Point().Pick(suite.XOF([]byte("H")))
-	n := 10
+	n := uint32(10)
 	t := 2*n/3 + 1
 	x := make([]kyber.Scalar, n) // trustee private keys
 	X := make([]kyber.Point, n)  // trustee public keys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		x[i] = suite.Scalar().Pick(suite.RandomStream())
 		X[i] = suite.Point().Mul(x[i], nil)
 	}
@@ -122,7 +122,7 @@ func TestPVSSDelete(test *testing.T) {
 
 	// (2) Share decryption (trustees)
 	sH := make([]kyber.Point, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		sH[i] = pubPoly.Eval(encShares[i].S.I).V
 	}
 
@@ -133,7 +133,7 @@ func TestPVSSDelete(test *testing.T) {
 	globalChallenge, err := computeGlobalChallenge(suite, len(X), pubPoly, encShares)
 	require.NoError(test, err)
 
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		if ds, err := DecShare(suite, H, X[i], sH[i], x[i], globalChallenge, encShares[i]); err == nil {
 			K = append(K, X[i])
 			E = append(E, encShares[i])
@@ -156,11 +156,11 @@ func TestPVSSDeleteFail(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
 	G := suite.Point().Base()
 	H := suite.Point().Pick(suite.XOF([]byte("H")))
-	n := 10
+	n := uint32(10)
 	t := 2*n/3 + 1
 	x := make([]kyber.Scalar, n) // trustee private keys
 	X := make([]kyber.Point, n)  // trustee public keys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		x[i] = suite.Scalar().Pick(suite.RandomStream())
 		X[i] = suite.Point().Mul(x[i], nil)
 	}
@@ -174,7 +174,7 @@ func TestPVSSDeleteFail(test *testing.T) {
 
 	// (2) Share decryption (trustees)
 	sH := make([]kyber.Point, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		sH[i] = pubPoly.Eval(encShares[i].S.I).V
 	}
 
@@ -182,10 +182,10 @@ func TestPVSSDeleteFail(test *testing.T) {
 	var E []*PubVerShare // good encrypted shares
 	var D []*PubVerShare // good decrypted shares
 
-	globalChallenge, err := computeGlobalChallenge(suite, n, pubPoly, encShares)
+	globalChallenge, err := computeGlobalChallenge(suite, int(n), pubPoly, encShares)
 	require.NoError(test, err)
 
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		if ds, err := DecShare(suite, H, X[i], sH[i], x[i], globalChallenge, encShares[i]); err == nil {
 			K = append(K, X[i])
 			E = append(E, encShares[i])
@@ -208,11 +208,11 @@ func TestPVSSBatch(test *testing.T) {
 	suite := edwards25519.NewBlakeSHA256Ed25519()
 	G := suite.Point().Base()
 	H := suite.Point().Pick(suite.XOF([]byte("H")))
-	n := 5
+	n := uint32(5)
 	t := 2*n/3 + 1
 	x := make([]kyber.Scalar, n) // trustee private keys
 	X := make([]kyber.Point, n)  // trustee public keys
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		x[i] = suite.Scalar().Pick(suite.RandomStream())
 		X[i] = suite.Point().Mul(x[i], nil)
 	}
@@ -233,7 +233,7 @@ func TestPVSSBatch(test *testing.T) {
 	sH0 := make([]kyber.Point, n)
 	sH1 := make([]kyber.Point, n)
 	sH2 := make([]kyber.Point, n)
-	for i := 0; i < n; i++ {
+	for i := uint32(0); i < n; i++ {
 		sH0[i] = p0.Eval(e0[i].S.I).V
 		sH1[i] = p1.Eval(e1[i].S.I).V
 		sH2[i] = p2.Eval(e2[i].S.I).V
@@ -266,11 +266,11 @@ func TestPVSSBatch(test *testing.T) {
 	Z3 := []*PubVerShare{E0[3], E1[3], E2[3]}
 
 	globalChallenges := make([]kyber.Scalar, 3)
-	globalChallenges[0], err = computeGlobalChallenge(suite, n, p0, e0)
+	globalChallenges[0], err = computeGlobalChallenge(suite, int(n), p0, e0)
 	require.NoError(test, err)
-	globalChallenges[1], err = computeGlobalChallenge(suite, n, p1, e1)
+	globalChallenges[1], err = computeGlobalChallenge(suite, int(n), p1, e1)
 	require.NoError(test, err)
-	globalChallenges[2], err = computeGlobalChallenge(suite, n, p2, e2)
+	globalChallenges[2], err = computeGlobalChallenge(suite, int(n), p2, e2)
 	require.NoError(test, err)
 
 	// (2) Share batch decryption (trustees)
